requirements.txt
firecrawl-py
openai
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv